pyproject.toml
[tool.poetry]
name = "re2g"
version = "0.1.0"
description = ""
authors = ["nelly <yongwoo@crowdworks.kr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9.0"
transformers = "^4.38.1"
torch = "^2.2.1"
black = "^24.2.0"
datasets = "^2.17.1"
matplotlib = "^3.8.3"
torchsummary = "^1.5.1"
pydantic-settings = "^2.2.1"
lightning = "^2.2.0.post0"
wandb = "^0.16.3"
python-dotenv = "^1.0.1"
langchain = "^0.1.11"
rank-bm25 = "^0.2.2"
chromadb = "^0.4.24"

[tool.poetry.group.dev.dependencies]
mock = "^5.1.0"
pytest = "^8.0.0"
pytest-env = "^1.1.3"
pytest-mock = "^3.12.0"
pytest-asyncio = "^0.23.4"
pytest-freezegun = "^0.4.2"
pytest-xdist = "^3.5.0"
pre-commit = "^3.2.0"
pytest-cov = "^4.1.0"
locust = "^2.23.1"
watchdog = "^4.0.0"
torchvision = "^0.17.1"
seaborn = "^0.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[black]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
pythonpath = ["."]
minversion = 6.0
addopts = ["-ra", "-q"]
testpaths = ["tests"]
